requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0 